django/requirements.txt
Django==5.2.5
mysqlclient==2.2.7
djangorestframework==3.16.1
django-cors-headers==4.7.0